Figured out the algorithm
I can actually do it in my head now. It's pretty cool. Now I need to
think of a way to simplify it so it runs faster, and then maybe figure
out a formula to solve and also what the numbers are supposed to be in
order to hit the right target. It's a little challenging.